simple_project/app/requirements.txt
asgiref==3.8.1
attrs==23.2.0
colorama==0.4.6
Django==5.1.6
django-bootstrap5==24.2
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-extensions==3.2.3
djangorestframework==3.15.1
et_xmlfile==2.0.0
Faker==25.0.1
flake8==5.0.4
flake8-docstrings==1.7.0
iniconfig==2.0.0
mccabe==0.7.0
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pep8==1.7.1
pep8-naming==0.13.3
pillow==10.4.0
pluggy==1.5.0
postgres==4.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.10
psycopg2-pool==1.2
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.3.0
pyflakes==2.5.0
pytest==7.1.3
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
pytest-subtests==0.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytils==0.4.1
pytz==2024.1
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uwsgi==2.0.20
split-settings==1.0.0